Prevent Tagged and Carted Enderman From Being Shulkerized
Adds a check that the enderman needs to be on the ground and not have a 
custom name. The check for enderman on ground is important to prevent 
enderman in minecrafts or boats from being shulkerized. The issue is the 
enderman can not be teleported to thier death while in a minecart.
That, and a player would likely notice thier enderman is no longer in 
thier minecart.